Upgrade Random forest + Update Client
Random Forest now using Entropy as the main method to calculate the cost when split node and COMPARISON as the main method to split the dataset by attribute values.
Random forest by default is generating 100 trees with boots trap data by random the data and buff the label 'B' data by 4 times.
Random forest has two method to create. First, it is 'build best model' and second is 'double training.
I have updated two more functions that help us control random forest. They are import and export function. In 'tree' class and 'randomForest' class, they now have functions that import data and export data from a string or from a file/ to a string or to a file.
The client has updated with some new things. It's a new UI that help us to manage things better and easier.
What a big update!